setup project
Move Draw2D into a subcrate to provide a clear separation between
the code being used by demos and the code required for rendering
graphics with vulkan.

The project is structured as a library with multiple binaries.
Binaries are located in the 'src/bin' directory and the library
is shared in the 'src' directory.

Note:  should draw2d be a git submodule?